+++ tooling/helium-server/webpack.config.js
@@ -1,9 +1,13 @@
+const path = require('path');
+
 module.exports = {
   entry: './worker/index.js',
   target: 'webworker',
   resolve: {
     mainFields: ['main', 'module'],
-    alias: {}
+    alias: {
+      tiConfig: path.resolve(__dirname, './ti-config.json')
+    }
   },
   node: { fs: 'empty' }
 };