+++ tooling/helium-server/worker/ssr.js
@@ -2,7 +2,8 @@ import { createPageRenderer } from 'vite-plugin-ssr';
 import jwt_decode from 'jwt-decode';
 // We load `importBuild.js` so that the worker code can be bundled into a single file
 import '../dist/server/importBuild.js';
-import { initPageContext, findTiInstance } from '@thoughtindustries/helium-server';
+import { initPageContext } from '@thoughtindustries/helium-server';
+import tiConfig from 'tiConfig';
 
 export { handleSsr };
 
@@ -36,7 +37,7 @@ const create =
 const sha256 = create('SHA-256');
 
 async function handleSsr(url, authToken = null) {
-  const tiInstance = await findTiInstance(INSTANCE_NAME);
+  const tiInstance = findTiInstance(INSTANCE_NAME);
   const { currentUser, appearanceBlock } = decryptUserAndAppearance(url, tiInstance);
   const pageContext = await initPageContext(
     url,
@@ -107,4 +108,18 @@ function assembleHeaders(pageContext) {
   }
 
   return headers;
+}
+
+function findTiInstance(instanceName) {
+  const { instances = [] } = tiConfig;
+  let instance = instances[0];
+
+  if (instanceName) {
+    const possibleMatch = instances.find(instance => instance.nickname === instanceName);
+    if (possibleMatch && possibleMatch.apiKey) {
+      instance = possibleMatch;
+    }
+  }
+
+  return instance;
 }