+++ tooling/create-helium-app/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.8](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.7...create-helium-app@1.0.0-alpha.8) (2022-02-15)
+
+
+### Bug Fixes
+
+* remove package-lock.json, add __npmrc file ([493b162](https://github.com/thoughtindustries/helium/commit/493b16254edcc8c5b707921465b4fdda9a2f21e3))
+
+
+
+
+
+# [1.0.0-alpha.7](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.6...create-helium-app@1.0.0-alpha.7) (2022-02-14)
+
+**Note:** Version bump only for package create-helium-app
+
+
+
+
+
+# [1.0.0-alpha.6](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.5...create-helium-app@1.0.0-alpha.6) (2022-02-14)
+
+
+### Bug Fixes
+
+* prepublish script ([7d75fae](https://github.com/thoughtindustries/helium/commit/7d75faec37eacb17a385c2772fbb721cd3f46502))
+* revert prepack script ([c5a9684](https://github.com/thoughtindustries/helium/commit/c5a9684921737c5bd013462e52eef566217e5a35))
+
+
+
+
+
+# [1.0.0-alpha.5](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.4...create-helium-app@1.0.0-alpha.5) (2022-02-14)
+
+
+### Bug Fixes
+
+* revert to preversion script ([1565c80](https://github.com/thoughtindustries/helium/commit/1565c802c6c8f31f98de723c4b90797e2741774b))
+
+
+
+
+
+# [1.0.0-alpha.4](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.3...create-helium-app@1.0.0-alpha.4) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix prepack script ([23769b6](https://github.com/thoughtindustries/helium/commit/23769b634e550f1525d1f34fe465fa8d53070927))
+
+
+
+
+
+# [1.0.0-alpha.3](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.2...create-helium-app@1.0.0-alpha.3) (2022-02-14)
+
+**Note:** Version bump only for package create-helium-app
+
+
+
+
+
+# [1.0.0-alpha.2](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.1...create-helium-app@1.0.0-alpha.2) (2022-02-14)
+
+**Note:** Version bump only for package create-helium-app
+
+
+
+
+
+# [1.0.0-alpha.1](https://github.com/thoughtindustries/helium/compare/create-helium-app@1.0.0-alpha.0...create-helium-app@1.0.0-alpha.1) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix script call for publish ([0a07257](https://github.com/thoughtindustries/helium/commit/0a07257bb64731706917dbadd0ffcc1e1f3099e5))
+
+
+
+
+
 ## [0.0.2-alpha.3](https://github.com/thoughtindustries/helium/compare/create-helium-app@0.0.2-alpha.2...create-helium-app@0.0.2-alpha.3) (2022-02-09)
 
 **Note:** Version bump only for package create-helium-app