+++ tooling/helium-server/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.0.3...@thoughtindustries/helium-server@1.0.4) (2022-02-15)
+
+
+### Bug Fixes
+
+* update webpack.config.js ([6355e0e](https://github.com/thoughtindustries/helium/commit/6355e0ec1ea8118d8231be37fe096dbbbb85d2f5))
+
+
+
+
+
+## [1.0.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.0.2...@thoughtindustries/helium-server@1.0.3) (2022-02-15)
+
+
+### Bug Fixes
+
+* make findTiInstance compatible with worker ([6fa91cb](https://github.com/thoughtindustries/helium/commit/6fa91cb933db9ca7ca5347b07b0f18a1b52faa55))
+
+
+
+
+
+## [1.0.2](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-server@1.0.1...@thoughtindustries/helium-server@1.0.2) (2022-02-14)
+
+**Note:** Version bump only for package @thoughtindustries/helium-server
+
+
+
+
+
 ## 1.0.1 (2022-02-14)
 
 **Note:** Version bump only for package @thoughtindustries/helium-server