+++ tooling/template-base/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0-alpha.8](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.7...@thoughtindustries/helium-template@1.1.0-alpha.8) (2022-02-15)
+
+
+### Bug Fixes
+
+* update webpack.config.js ([6355e0e](https://github.com/thoughtindustries/helium/commit/6355e0ec1ea8118d8231be37fe096dbbbb85d2f5))
+
+
+
+
+
+# [1.1.0-alpha.7](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.6...@thoughtindustries/helium-template@1.1.0-alpha.7) (2022-02-15)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template
+
+
+
+
+
+# [1.1.0-alpha.6](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.5...@thoughtindustries/helium-template@1.1.0-alpha.6) (2022-02-15)
+
+
+### Bug Fixes
+
+* remove package-lock.json, add __npmrc file ([493b162](https://github.com/thoughtindustries/helium/commit/493b16254edcc8c5b707921465b4fdda9a2f21e3))
+
+
+
+
+
+# [1.1.0-alpha.5](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.4...@thoughtindustries/helium-template@1.1.0-alpha.5) (2022-02-14)
+
+**Note:** Version bump only for package @thoughtindustries/helium-template
+
+
+
+
+
+# [1.1.0-alpha.4](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.3...@thoughtindustries/helium-template@1.1.0-alpha.4) (2022-02-14)
+
+
+### Bug Fixes
+
+* template-base publishConfig ([ab3a0dd](https://github.com/thoughtindustries/helium/commit/ab3a0ddfce4693607622be5770327eff492f31d0))
+
+
+
+
+
+# [1.1.0-alpha.3](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.2...@thoughtindustries/helium-template@1.1.0-alpha.3) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix template-base privacy ([1ba6bdd](https://github.com/thoughtindustries/helium/commit/1ba6bdd8cb1293dfd2fa5b391e2f3f4c84effca8))
+
+
+
+
+
 # [1.1.0-alpha.2](https://github.com/thoughtindustries/helium/compare/@thoughtindustries/helium-template@1.1.0-alpha.1...@thoughtindustries/helium-template@1.1.0-alpha.2) (2022-02-14)
 
 **Note:** Version bump only for package @thoughtindustries/helium-template